+++ src/Enumeration.php
@@ -3,6 +3,7 @@
 namespace Sourceboat\Enumeration;
 
 use Eloquent\Enumeration\AbstractEnumeration;
+use Illuminate\Support\Str;
 use Sourceboat\Enumeration\Rules\EnumerationValue;
 
 /**
@@ -37,6 +38,17 @@ abstract class Enumeration extends AbstractEnumeration
     public function localized(): string
     {
         return trans(static::getLocalizationPath() . '.' . $this->value());
+    }
+
+    /**
+     * Check if this instance equals to a specific member of the enum.
+     *
+     * @param static $value The member to check for
+     * @return bool
+     */
+    public function is($value): bool
+    {
+        return $this === $value;
     }
 
     /**
@@ -168,10 +180,25 @@ abstract class Enumeration extends AbstractEnumeration
      * Returns a string representation of this member.
      *
      * @return string The string representation.
-     * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingReturnTypeHint
      */
-    public function __toString()
+    public function __toString(): string
     {
         return (string) $this->value();
+    }
+
+    /**
+     * Implements `is<EnumKey>` methods for the enum.
+     *
+     * @param string $method
+     * @param array $arguments
+     * @return mixed
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.TypeHintDeclaration.MissingParameterTypeHint
+     */
+    public function __call($method, $arguments)
+    {
+        if (Str::startsWith($method, 'is')) {
+            $key = Str::after($method, 'is');
+            return $this->is(static::memberByKey($key, false));
+        }
     }
 }